+++ FE/src/components/FeedCard/CommentCreate.js
@@ -1,0 +1,58 @@
+import { styled } from "styled-components";
+
+const CommentCreate = () => {
+  return (
+    <>
+      <CreateSection>
+        <Label for="CommentInput">ㅇ</Label>
+        <Input
+          id="CommentInput"
+          placeholder="여기에 댓글을 입력해주세요."
+        ></Input>
+        <CreateButton>등록</CreateButton>
+      </CreateSection>
+    </>
+  );
+};
+
+export default CommentCreate;
+
+const CreateSection = styled.section`
+  width: 90%;
+
+  margin-bottom: 10px;
+
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+`;
+
+const Label = styled.label`
+  width: 20px;
+  height: 20px;
+`;
+
+const Input = styled.input`
+  width: 220px;
+  height: 30px;
+
+  margin: 0px 15px;
+  border-bottom: 1px solid #d0d0d0;
+`;
+
+const CreateButton = styled.button`
+  font-size: 0.75rem;
+
+  background-color: #ececec;
+  &:hover {
+    background-color: #d0d0d0;
+  }
+
+  padding: 5px 15px;
+  border-radius: 15px;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
