+++ FE/src/data/dummy.js
@@ -38,162 +38,162 @@ const todoList = {
     {
       todoId: 1,
       memberId: 1,
-      title: "투두1",
-      tag: "태그1",
-      emoji: "⏰",
+      content: "투두1",
+      tagName: "태그1",
+      todoEmoji: "⏰",
       complete: true,
       date: "2023-08-25",
     },
     {
       todoId: 2,
       memberId: 1,
-      title: "투두2",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두2",
+      tagName: "태그2",
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 3,
       memberId: 1,
-      title: "투두3",
-      tag: "태그3",
-      emoji: "⏰",
+      content: "투두3",
+      tagName: "태그3",
+      todoEmoji: "⏰",
       complete: true,
       date: "2023-08-25",
     },
     {
       todoId: 4,
       memberId: 1,
-      title: "투두4",
-      tag: "태그4",
-      emoji: "⏰",
+      content: "투두4",
+      tagName: "태그4",
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 5,
       memberId: 1,
-      title: "투두5",
-      tag: "태그3",
-      emoji: "⏰",
+      content: "투두5",
+      tagName: "태그3",
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 6,
       memberId: 1,
-      title: "투두6",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두6",
+      tagName: "태그2",
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 7,
       memberId: 1,
-      title: "투두7",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두7",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 8,
       memberId: 1,
-      title: "투두7",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두7",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 9,
       memberId: 1,
-      title: "투두7",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두7",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-25",
     },
     {
       todoId: 1,
       memberId: 1,
-      title: "투두1",
+      content: "투두1",
       tag: "태그1",
-      emoji: "⏰",
+      todoEmoji: "⏰",
       complete: true,
       date: "2023-08-21",
     },
     {
       todoId: 2,
       memberId: 1,
-      title: "투두2",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두2",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-23",
     },
     {
       todoId: 3,
       memberId: 1,
-      title: "투두3",
+      content: "투두3",
       tag: "태그3",
-      emoji: "⏰",
+      todoEmoji: "⏰",
       complete: true,
       date: "2023-08-24",
     },
     {
       todoId: 4,
       memberId: 1,
-      title: "투두4",
+      content: "투두4",
       tag: "태그4",
-      emoji: "⏰",
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-22",
     },
     {
       todoId: 5,
       memberId: 1,
-      title: "투두5",
+      content: "투두5",
       tag: "태그3",
-      emoji: "⏰",
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-21",
     },
     {
       todoId: 6,
       memberId: 1,
-      title: "투두6",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두6",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-22",
     },
     {
       todoId: 7,
       memberId: 1,
-      title: "투두7",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두7",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-23",
     },
     {
       todoId: 8,
       memberId: 1,
-      title: "투두7",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두7",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-24",
     },
     {
       todoId: 9,
       memberId: 1,
-      title: "투두7",
-      tag: "태그2",
-      emoji: "⏰",
+      content: "투두7",
+      tagResponse: { tagName: "태그2" },
+      todoEmoji: "⏰",
       complete: false,
       date: "2023-08-24",
     },
@@ -210,10 +210,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "2023년 08월 25일 11시 11분",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: true },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: true },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: true,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: true,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: true,
       likes: 2021,
@@ -287,10 +307,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "2023년 08월 25일 11시 11분",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: false,
       likes: 2021,
@@ -323,10 +363,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "08.25 11:11",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: false,
       likes: 2021,
@@ -358,10 +418,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "08.25 11:11",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: false,
       likes: 2021,
@@ -393,10 +473,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "08.25 11:11",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: false,
       likes: 2021,
@@ -428,10 +528,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "08.25 11:11",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: false,
       likes: 2021,
@@ -463,10 +583,30 @@ const postList = {
       imgUrl: "FE/src/assets/images/memberImg.jpeg",
       createdAt: "08.25 11:11",
       todoList: [
-        { title: "투두6", tag: "태그2", emoji: "🐣", complete: true },
-        { title: "투두7", tag: "태그2", emoji: "🐣", complete: true },
-        { title: "투두8", tag: "태그2", emoji: "🐣", complete: false },
-        { title: "투두9", tag: "태그2", emoji: "🐣", complete: false },
+        {
+          content: "투두6",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: true,
+        },
+        {
+          content: "투두7",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: true,
+        },
+        {
+          content: "투두8",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
+        {
+          content: "투두9",
+          tagResponse: { tagName: "태그2" },
+          todoEmoji: "🐣",
+          complete: false,
+        },
       ],
       like: false,
       likes: 2021,
