+++ FE/src/pages/todoPage.js
@@ -20,6 +20,7 @@ import "react-calendar/dist/Calendar.css"; // css import
 
 //삭제 할 더미 데이터들
 import trophyLevel1 from "../assets/images/trophyLevel1.png";
+import TodoCard from "../components/TodoCard";
 const membersId = 1;
 
 const TodoPage = () => {
@@ -272,29 +273,6 @@ const TodoModalBody = styled.div`
   align-items: center;
 `;
 
-const TodoSection = styled.section`
-  width: 300px;
-  height: 150px;
-
-  margin-bottom: 30px;
-  border: 1px solid #ffd900;
-  border-radius: 15px;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  align-items: end;
-`;
-
-const ElDiv = styled.div`
-  width: 300px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-`;
-
 const Exit = ({ ExitTodoModal }) => {
   return (
     <ExitBody>
@@ -355,6 +333,62 @@ const Todo = ({ todo }) => {
     </>
   );
 };
+
+const TodoSection = styled.section`
+  width: 300px;
+  height: 150px;
+
+  margin-bottom: 30px;
+  border: 1px solid #ffd900;
+  border-radius: 15px;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  align-items: end;
+`;
+
+const ElDiv = styled.div`
+  width: 300px;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+const TagDiv = styled.span`
+  width: 70px;
+  height: 40px;
+
+  background: #ececec;
+
+  margin: 15px;
+  border-radius: 15px;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+const TitleDiv = styled.p`
+  width: 200px;
+
+  font-size: 0.9erm;
+`;
+
+const EmojiDiv = styled.div`
+  width: 40px;
+  height: 40px;
+
+  margin: 15px;
+  border: 1px solid #949597;
+  border-radius: 15px;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
 
 const TodoModalButtons = ({ todo, ChangeComplete, navigate, DeleteTodo }) => {
   return (
@@ -628,7 +662,12 @@ const TodoList = ({ todos, ChangeTodo }) => {
       <ListContainer>
         <Todos>
           {todos.map((value, idx) => (
-            <TodoCard key={idx} value={value} ChangeTodo={ChangeTodo} />
+            <TodoCard
+              key={idx}
+              value={value}
+              todoPage={true}
+              ChangeTodo={ChangeTodo}
+            />
           ))}
         </Todos>
       </ListContainer>
@@ -656,65 +695,5 @@ const Todos = styled.ul`
 
   display: flex;
   flex-direction: column;
-  align-items: center;
-`;
-
-const TodoCard = ({ value, ChangeTodo }) => {
-  return (
-    <>
-      <TodoCardSection onClick={() => ChangeTodo(value)}>
-        <TagDiv>{value.tagResponse.tagName}</TagDiv>
-        <TitleDiv>{value.content}</TitleDiv>
-        <EmojiDiv>{value.complete === "DONE" ? value.todoEmoji : ""}</EmojiDiv>
-      </TodoCardSection>
-    </>
-  );
-};
-
-const TodoCardSection = styled.li`
-  width: 390px;
-  height: 70px;
-
-  margin-top: 20px;
-  border-radius: 15px;
-
-  background-color: #ffffff;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const TagDiv = styled.span`
-  width: 70px;
-  height: 40px;
-
-  background: #ececec;
-
-  margin: 15px;
-  border-radius: 15px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const TitleDiv = styled.p`
-  width: 200px;
-
-  font-size: 0.9erm;
-`;
-
-const EmojiDiv = styled.div`
-  width: 40px;
-  height: 40px;
-
-  margin: 15px;
-  border: 1px solid #949597;
-  border-radius: 15px;
-
-  display: flex;
-  justify-content: center;
   align-items: center;
 `;
